MarkerBackend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.4
fastapi==0.100.0
h11==0.14.0
httptools==0.6.0
idna==3.4
pydantic==2.0.2
pydantic_core==2.1.2
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
python-socketio